Adjust frame’s number field.
Closes #269

https://adamj.eu/tech/2021/05/08/django-check-constraints-limit-range-integerfield/

+++ inventory/models.py
@@ -512,7 +512,7 @@ class Journal(models.Model):
 
 class Frame(models.Model):
     roll = models.ForeignKey(Roll, on_delete=models.CASCADE)
-    number = models.IntegerField()
+    number = models.PositiveSmallIntegerField()
     date = models.DateField(default=datetime.date.today)
     notes = models.TextField(blank=True)
     aperture = models.CharField(
